# public/openapi.yaml
openapi: 3.0.3
info:
  title: Edusm User & Admin Management API
  version: 1.0.0
  description: API for creating users/admins, login, and role-based validation with token and superadmin_email.
servers:
  - url: https://next-js-online-learning-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /admin/create:
    get:
      summary: Get all users
      operationId: GetCreateUser
      tags:
        - Admin - User Management
      description: Fetches a list of all users.
      responses:
        '200':
          description: Successfully fetched users
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Users fetched successfully"
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: integer
                          example: 1
                        email:
                          type: string
                          example: "user@example.com"
                        role:
                          type: string
                          example: "user"
                        active_status:
                          type: boolean
                          example: true
        '500':
          description: Failed to fetch users
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch users"
                  details:
                    type: string
                    example: "Database connection error"

    post:
      summary: Create a new user or admin
      operationId: PostCreateUser
      tags:
        - Admin - User Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                token:
                  type: string
                  description: Required if superadmin_email is present
                superadmin_email:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '400':
          description: Validation error (missing fields)
        '403':
          description: Unauthorized (invalid token or superadmin_email)
        '409':
          description: Conflict (email already exists)

    put:
      summary: Update user details
      operationId: UpdatePutUser
      tags:
        - Admin - User Management
      description: Updates the email, password, role, or active status of a specific user.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
            example: 1
          description: The ID of the user to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "updateduser@example.com"
                password:
                  type: string
                  example: "newpassword"
                role:
                  type: string
                  example: "admin"
                active_status:
                  type: boolean
                  example: true
                token:
                  type: string
                  description: "Token for superadmin verification"
                  example: "superadmintoken"
                superadmin_email:
                  type: string
                  description: "Email of the superadmin"
                  example: "superadmin@example.com"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated successfully"
        '400':
          description: Missing required fields or no fields to update
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email and password are required"
        '403':
          description: Invalid token or superadmin email
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired token for superadmin"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found"
        '500':
          description: Failed to update user
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to update user"
                  details:
                    type: string
                    example: "Database query error"
  /admin/login:
    post:
      summary: Login admin with email and password
      tags:
        - Admin - User Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
                  email:
                    type: string
        '401':
          description: Invalid credentials
        '404':
          description: User not found
...
  /blog/category:
    get:
      summary: Fetch all blog categories
      operationId: BlogCategory
      tags:
        - Blog
      responses:
        '200':
          description: A list of blog categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cat_id:
                      type: integer
                    cat_name:
                      type: string
        '500':
          description: Server error

    post:
      summary: Add a new blog category
      operationId: AddBlogCategory
      tags:
        - Blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cat_name:
                  type: string
              required:
                - cat_name
      responses:
        '200':
          description: Category added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          description: Category already exists
        '500':
          description: Error adding category

    put:
      summary: Update an existing blog category
      operationId: UpdateBlogCategory
      tags:
        - Blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cat_id:
                  type: integer
                cat_name:
                  type: string
              required:
                - cat_id
      responses:
        '200':
          description: Category updated successfully
        '400':
          description: Bad request
        '404':
          description: Category not found
        '500':
          description: Error updating category

    delete:
      summary: Delete a blog category
      operationId: DeleteBlogCategory
      tags:
        - Blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cat_id:
                  type: integer
              required:
                - cat_id
      responses:
        '200':
          description: Category deleted
        '500':
          description: Error deleting category